edgefleet-prototype/backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Data Processing & Analysis
pandas==2.1.3
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# API & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment Variables
python-dotenv==1.0.0

# HTTP Client
aiohttp==3.9.1
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
loguru==0.7.2

# Computer Vision
opencv-python-headless==4.8.1.78
torch==2.1.0
torchvision==0.16.0
numpy>=1.21.0

# Image Processing
Pillow==10.1.0

# CORS Middleware (included in FastAPI)
# Additional utilities
typing-extensions==4.8.0